requirements.txt
pandas
numpy
matplotlib 
plotly.express
random 
string 
datetime